+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Saved translation files now end in a newline
+- Translations removed from the source will now be removed from all destinations
 
 ### Changed